Make work in Azure only environment by using ipaddress instead of FQDN
The FQDN for Azure is a full internal path that can not be used between the machines for SSH access.
Using the ipaddress will work for our needs, but a more general solution should be used.
The bootstrap user will change shortly by pulling in a request made https://github.com/oswaldlabs/chef-rundeck/pull/31